+++ config/settings/test.py
@@ -124,6 +124,7 @@ SHEETS_EXTERNAL_LINK_PATH = os.path.join(
 SHEETS_IGV_OMIT_BAM = ['*dragen_evidence.bam']
 SHEETS_IGV_OMIT_VCF = ['*cnv.vcf.gz', '*ploidy.vcf.gz', '*sv.vcf.gz']
 SHEETS_API_FILE_EXISTS_RESTRICT = False
+SHEETS_PARSER_WARNING_SAVE_LIMIT = 100
 
 # Landingzones app settings
 LANDINGZONES_TRIGGER_ENABLE = True